Alright, let's see what we got here
"Commit Message"

First and foremost, spotlessapply ofc
Side note most of these changes are shit I did BEFORE vanguard fried itself so idfr what half of these are I just recovered them from my SSD's

Looks like I GREATLY improved flightloopsound, altho I remember them still sucking ASS, so flight sounds need testing.

Moved a buncha ResourceLocation shit over to UniversalCommon to help with porting to diff MC versions

Seperated a shit tonna ticking stuff, TickingTile now has a ClientTick and a Tick, one runs only on client, the other only on server. Obviously.

Fixed the advanced control circuit model/animation

THE FABRICATOR! HA!

A new rendertype for arcs? that's for like electricity rendering shits

textured the basic nozzle

Had to majorly fuck with the build.gradle to get rust working again on my system, praying it works on everyone elses + github

Started work on ComputerCraft integration!

Messed with... a LOT of planet-related stuff ig

Removed useless ass chromium

+++ src/main/java/com/code/tama/tts/server/data/json/dataHolders/DataDimOxygen.java
@@ -7,7 +7,7 @@ import net.minecraft.resources.ResourceLocation;
 
 @Builder
 public class DataDimOxygen {
-	public float oxygen;
+	public int oxygen;
 	public ResourceLocation dimension;
 
 	@Override
